Move schema back into commands dir
